Refactor

+++ client/commands/react.js
@@ -7,8 +7,7 @@ module.exports.run = async (client, msg, args) => {
     let emoji = await client.emojis.find(emoji => emoji.name === args[0]) || client.emojis.get(args[0]);
     if (!emoji) return msg.reply("Couldn't find emoji.");
     await msg.channel.send(`${emoji}`);
-    if (!msg.channel.permissionsFor(msg.guild.me).has("MANAGE_MESSAGES")) return;
-    await msg.delete();
+    if (msg.deletable) await msg.delete();
 
 };
